pyproject.toml
[project]
name = "django_tools"
version = "0.4.4"
description = "Django plugin with Pydantic settings, Celery integration, and Dependency Injection"
readme = "README.md"
authors = [
    { name = "Eduardo Jesus", email = "dudulj15@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "redis>=5.0.0",
    "dependency-injector>=4.48.2",
    "kombu>=5.5.4",
    "orjson>=3.11.4",
    "django-stubs>=5.2.7",
    "django-celery-results>=2.6.0",
    "django-ninja>=1.4.5",
]



[project.urls]
Homepage = "https://github.com/edududs/django_tools"
Repository = "https://github.com/edududs/django_tools"
Issues = "https://github.com/edududs/django_tools/issues"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[tool.setuptools.packages.find]
where = ["src"]
include = ["django_tools*"]

# Configuração do Ruff baseada no seu VS Code
[tool.ruff]
target-version = "py311"
line-length = 120
fix = true
show-fixes = true

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=9.6.0",
    "pytest>=8.4.2",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.14.0",
    "rich>=14.2.0",
    "ruff>=0.8.0",
    "typer>=0.20.0",
    "types-redis>=4.0.0",
]
    


